merger main and config
