+++ todoLy-api/build.gradle
@@ -1,0 +1,46 @@
+plugins {
+    id 'java'
+}
+
+dependencies {
+    implementation project(":todoLy-core")
+    // https://mvnrepository.com/artifact/io.rest-assured/rest-assured
+    implementation group: 'io.rest-assured', name: 'rest-assured', version: '5.5.0'
+    // https://mvnrepository.com/artifact/io.rest-assured/json-schema-validator
+    implementation group: 'io.rest-assured', name: 'json-schema-validator', version: '5.5.0'
+
+}
+
+test {
+    useTestNG();
+}
+// Option 1 After
+tasks.register("executeApiFeature", Test) {
+//    systemProperties(System.getProperties())
+    outputs.upToDateWhen { false }
+    useTestNG()
+    scanForTestClasses = false
+    options {
+        useDefaultListeners = false
+        jvmArgs = ["-Dcucumber.features=${rootProject.rootDir}/trello-api/src/test/resources/features/"]
+        jvmArgs = ["-Dcucumber.plugin=pretty, io.qameta.allure.cucumber7jvm.AllureCucumber7Jvm"]
+        jvmArgs = ["-Dcucumber.filter.tags=" + tags]
+        jvmArgs = ["-Denv=" + env]
+        testLogging.showStandardStreams = true
+    }
+}
+// Option 2 Before Module for movile, UI
+task apiFeature(type: Test) {
+    useTestNG()
+    options {
+        jvmArgs = ['-Dcucumber.execution.strict=' + true]
+//        jvmArgs = ['-Dcucumber.plugin=io.qameta.allure.cucumber7jvm.AllureCucumber7Jvm']
+//        jvmArgs = ["-Dcucumber.filter.tags=" + tags]
+        jvmArgs = ["-Denv=" + env]
+    }
+    //Flag to show console output
+    testLogging.showStandardStreams = true
+    //Flag to scan for classes with @Test annotation.
+    scanForTestClasses = false
+    // finalizedBy 'allureReport'
+}
