+++ todoLy-api/src/test/resources/schemas/deleteBoardSchema.json
