--- trello-core/build.gradle
@@ -1,24 +1,0 @@
-plugins {
-    id 'java'
-}
-
-group = 'com.trello'
-version = '1.0'
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    // https://mvnrepository.com/artifact/org.apache.logging.log4j/log4j-core
-    implementation group: 'org.apache.logging.log4j', name: 'log4j-core', version: '2.20.0'
-    // https://mvnrepository.com/artifact/com.jayway.jsonpath/json-path
-    implementation group: 'com.jayway.jsonpath', name: 'json-path', version: '2.9.0'
-    // https://mvnrepository.com/artifact/com.googlecode.json-simple/json-simple
-    implementation group: 'com.googlecode.json-simple', name: 'json-simple', version: '1.1.1'
-
-}
-
-test {
-    useTestNG()
-}
